always use int values for area combinations
- rename AreaDictionaryShort to AreaDictionary
- remove AreaDictionaryInt.java and the SparseLong2ShortMap sources

pro: nearly no practical limits reg. number of areas
contra: a bit slower 

Memory consumption is not yet fully optimized

git-svn-id: http://svn.mkgmap.org.uk/splitter/branches/refactoring2@510 838fabe5-b1a8-4b81-9896-7a84386980e9

+++ src/uk/me/parabola/splitter/AbstractMapProcessor.java
@@ -16,7 +16,7 @@ package uk.me.parabola.splitter;
 import java.util.concurrent.BlockingQueue;
 
 public abstract class AbstractMapProcessor implements MapProcessor {
-	public static final short UNASSIGNED = Short.MIN_VALUE;
+	public static final int UNASSIGNED = Short.MIN_VALUE;
 
 	public boolean skipTags(){
 		return false;
